sd_combined: fix SparseBox3DKeyPointsGenerator dtype
torch.tensor(fix_scale) infers Long when all entries are int (e.g.
fix_scale=[[0, 0, 0]]). nn.Parameter then holds a Long tensor and
mmcv's init_weights crashes on param.data.mean(). Force float32.

Repro: B1.5 (3398019), 2a-noaux (3398319), 2a-aux (3398320 cancelled)
all crashed at init with "mean(): could not infer output dtype".

+++ projects/mmdet3d_plugin/models/detection3d/detection3d_blocks.py
@@ -199,7 +199,7 @@ class SparseBox3DKeyPointsGenerator(BaseModule):
         if fix_scale is None:
             fix_scale = ((0.0, 0.0, 0.0),)
         self.fix_scale = nn.Parameter(
-            torch.tensor(fix_scale), requires_grad=False
+            torch.tensor(fix_scale, dtype=torch.float32), requires_grad=False
         )
         self.num_pts = len(self.fix_scale) + num_learnable_pts
         if num_learnable_pts > 0:
